LayoutTests/platform/glib/imported/w3c/web-platform-tests/cookies/samesite/iframe-reload.https-expected.txt

CONSOLE MESSAGE: [object MessageEvent]
CONSOLE MESSAGE: [object MessageEvent]
CONSOLE MESSAGE: [object MessageEvent]

PASS Reloaded same-host fetches are strictly same-site
PASS Reloaded subdomain fetches are strictly same-site
PASS Reloaded cross-site fetches are cross-site
